move homepage out and accessible on logged out and logged in state
