+++ frontend/src/pages/Home.tsx
@@ -1,237 +1,199 @@
-import React, { FC, useState, useCallback } from 'react'
-import axios from 'axios'
+import { FC } from 'react'
 import Button from '@mui/material/Button'
-import TextField from '@mui/material/TextField'
-import Typography from '@mui/material/Typography'
-import Grid from '@mui/material/Grid'
-import Container from '@mui/material/Container'
-import Box from '@mui/material/Box'
-import MenuItem from '@mui/material/MenuItem'
-import Select from '@mui/material/Select'
-import InputLabel from '@mui/material/InputLabel'
-import FormControl from '@mui/material/FormControl'
-import useFilestore from '../hooks/useFilestore'
-import FileUpload from '../components/widgets/FileUpload'
-import CloudUploadIcon from '@mui/icons-material/CloudUpload'
-import useSnackbar from '../hooks/useSnackbar'
-import useApi from '../hooks/useApi'
-import useRouter from '../hooks/useRouter'
-import useAccount from '../hooks/useAccount'
-import Homepage from './Homepage'
+import { styled } from '@mui/system';
 
-const Dashboard: FC = () => {
-  const filestore = useFilestore()
-  const snackbar = useSnackbar()
-  const api = useApi()
-  const {navigate} = useRouter()
-  const account = useAccount()
+const XContainer = styled('div')({
+    maxWidth: '800px',
+    margin: '0 auto',
+    padding: '20px',
+});
 
-  const [loading, setLoading] = useState(false)
-  const [inputValue, setInputValue] = useState('')
-  const [chatHistory, setChatHistory] = useState<Array<{user: string, message: string}>>([])
-  const [selectedMode, setSelectedMode] = useState('Create')
-  const [selectedCreateType, setSelectedCreateType] = useState('Text')
-  const [selectedFineTuneType, setSelectedFineTuneType] = useState('Text')
-  const [files, setFiles] = useState<File[]>([])
+const Header = styled('div')({
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center',
+    height: '100px',
+});
 
-  const handleInputChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setInputValue(event.target.value)
-  }
+const Block = styled('div')({
+    display: 'flex',
+    alignItems: 'center',
+    padding: '40px 20px',
+    marginBottom: '40px',
+    // boxShadow: '0 4px 8px rgba(0,0,0,0.1)',
+});
 
-  const onSend = async () => {
-      // const statusResult = await axios.post('/api/v1/sessions', {
-      //   files: files,
-      // })
-      try {
-        const formData = new FormData()
-        files.forEach((file) => {
-          formData.append("files", file)
-        })
+const RightMedia = styled('div')({
+    flex: '1',
+    // paddingRight: '39px',
+});
 
-        formData.set('input', inputValue)
-        formData.set('mode', selectedMode)
-        if (selectedMode == "Create") {
-          formData.set("type", selectedCreateType)
-        } else {
-          formData.set("type", selectedFineTuneType)
-        }
+const RightContent = styled('div')({
+    flex: '1',
+    textAlign: 'left',
+    fontWeight: 499,
+});
 
-        await api.post('/api/v1/sessions', formData, {
-          // params: {
-          //   path,
-          // },
-          // onUploadProgress: (progressEvent) => {
-          //   const percent = progressEvent.total && progressEvent.total > 0 ?
-          //     Math.round((progressEvent.loaded * 100) / progressEvent.total) :
-          //     0
-          //   setUploadProgress({
-          //     percent,
-          //     totalBytes: progressEvent.total || 0,
-          //     uploadedBytes: progressEvent.loaded || 0,
-          //   })
-          // }
-        }).then((response) => {
-          account.loadSessions()
-          setFiles([])
-          setInputValue("")
-          console.log("ABOUT TO NAVIGATE")
-          navigate('session', {session_id: response.id})
-          console.log("DONE NAVIGATE")
-        })
-        // result = true
-      } catch(e) {
-        console.log(e)
-      }
-      // setUploadProgress(undefined)
-      // return result
+const LeftMedia = styled('div')({
+    flex: '1',
+    paddingRight: '39px',
+});
 
-    // TODO: put this in state, when user clicks send, POST all three things
-    // (files, text, type) to a new endpoint which accepts files
+const LeftContent = styled('div')({
+    flex: '1',
+    textAlign: 'left',
+    fontWeight: 499,
+    paddingRight: '39px',
+});
 
-    // const result = await filestore.upload("lhwoo", files)
-    // if(!result) return
-    // await filestore.loadFiles(filestore.path)
-    // snackbar.success('Files Uploaded')
-  }
+function OpenAIBlock() {
+    return (
+        <Block>
+            <LeftContent>
+                <img src="/img/helix-text-logo.png" alt="Helix Logo" style={{width:"250px"}} />
+                <h1>Open AI 😉</h1>
+                <p>Deploy the latest open source models securely in your cloud</p>
+                <p>Or let us run them for you</p>
+            </LeftContent>
+            <RightMedia>
+                <video autoPlay loop muted style={{width:"350px", float: "right", marginRight: "-50px"}}>
+                    <source src="/img/typing.mp4" type="video/mp4"/>
+                </video>
+            </RightMedia>
+        </Block>
+    );
+}
 
-  const onUpload = useCallback(async (files: File[]) => {
-    console.log(files)
-    setFiles(files)
-  }, [
-    filestore.path,
-  ])
-
-  const handleKeyDown = (event: React.KeyboardEvent<HTMLDivElement>) => {
-    if (event.key === 'Enter' && (event.shiftKey || event.ctrlKey)) {
-      onSend()
-      event.preventDefault()
-    }
-  }
-
-  return account.user ? (
-    <Container sx={{ mt: 4, mb: 4, display: 'flex', flexDirection: 'column', justifyContent: 'flex-end', overflowX: 'hidden' }}>
-      <Grid container spacing={3} direction="row" justifyContent="flex-start">
-        <Grid item xs={2} md={2}>
-        </Grid>
-        <Grid item xs={4} md={4}>
-          <Button variant={selectedMode === 'Create' ? "contained" : "outlined"} color="primary" sx={{ borderRadius: 35, mr: 2 }} onClick={() => setSelectedMode('Create')}>
-            Create
-            <FormControl sx={{ minWidth: 120, marginLeft: 2 }}>
-              <Select variant="standard"
-                labelId="create-type-select-label"
-                id="create-type-select"
-                value={selectedCreateType}
-                onChange={(event) => setSelectedCreateType(event.target.value)}
-              >
-                <MenuItem value="Text">Text</MenuItem>
-                <MenuItem value="Image">Images</MenuItem>
-              </Select>
-            </FormControl>
-          </Button>
-        </Grid>
-        <Grid item xs={4} md={4}>
-          <Button variant={selectedMode === 'Finetune' ? "contained" : "outlined"} color="primary" sx={{ borderRadius: 35, mr: 2 }} onClick={() => setSelectedMode('Finetune')}>
-            Finetune
-            <FormControl sx={{minWidth: 120, marginLeft: 2}}>
-              <Select variant="standard"
-                labelId="fine-tune-type-select-label"
-                id="fine-tune-type-select"
-                value={selectedFineTuneType}
-                onChange={(event) => setSelectedFineTuneType(event.target.value)}
-              >
-                <MenuItem value="Text">Text</MenuItem>
-                <MenuItem value="Image">Images</MenuItem>
-              </Select>
-            </FormControl>
-          </Button>
-        </Grid>
-        <Grid item xs={2} md={2}>
-        </Grid>
-        <Grid item xs={12} md={12}>
-          {chatHistory.map((chat, index) => (
-            <Typography key={index}><strong>{chat.user}:</strong> {chat.message}</Typography>
-          ))}
-        </Grid>
-      </Grid>
-      <Grid container item xs={12} md={8} direction="row" justifyContent="space-between" alignItems="center" sx={{ mt: 'auto', position: 'absolute', bottom: '5em', maxWidth: '800px' }}>
-        <Grid item xs={12} md={11}>
-          {selectedMode === 'Finetune' && selectedFineTuneType === 'Image' && (
-            <FileUpload
-              sx={{
-                width: '100%',
-                mt: 2,
-              }}
-              onUpload={ onUpload }
-            >
-              <Button
-                sx={{
-                  width: '100%',
-                }}
-                variant="contained"
-                color="secondary"
-                endIcon={<CloudUploadIcon />}
-              >
-                Upload Files
-              </Button>
-              <Box
-                sx={{
-                  border: '1px dashed #ccc',
-                  p: 2,
-                  display: 'flex',
-                  flexDirection: 'row',
-                  alignItems: 'center',
-                  justifyContent: 'center',
-                  minHeight: '100px',
-                  cursor: 'pointer',
-                  mb: 2,
-                }}
-              >
-                <Typography
-                  sx={{
-                    color: '#999'
+function ImageModelsBlock() {
+    return (
+        <Block>
+            <LeftMedia>
+                <img src="/img/sdxl.png" alt="Stable Diffusion XL" style={{width:"300px"}} />
+            </LeftMedia>
+            <RightContent>
+                <h1>Image models</h1>
+                <p>Train your own SDXL customized to your brand or style</p>
+                <Button
+                  variant="contained"
+                  onClick={ () => {
+                    // endIcon={<LoginIcon />}
+                    // account.onLogin()
                   }}
-                  variant="caption"
-                >
-                  drop files here to upload them ...
-                  {
-                    files.length > 0 && files.map((file) => (
-                      <Typography key={file.name}>
-                        {file.name} ({file.size} bytes) - {file.type}
-                      </Typography>
-                    ))
-                  }
-                </Typography>
-              </Box>
-            </FileUpload> )}
-        </Grid>
-        <Grid item xs={12} md={11}>
-            <TextField
-              fullWidth
-              label={(
-                selectedMode === 'Create' && selectedCreateType === 'Text' ? 'Start a chat with a base Mistral-7B-Instruct model' : selectedMode === 'Create' && selectedCreateType === 'Image' ? 'Describe an image to create it with a base SDXL model' : selectedMode === 'Finetune' && selectedFineTuneType === 'Text' ? 'Enter question-answer pairs to fine tune a language model' : 'Upload images and label them to fine tune an image model'
-                ) + " (shift+enter to send)"
-              }
-              value={inputValue}
-              disabled={loading}
-              onChange={handleInputChange}
-              name="ai_submit"
-              multiline={true}
-              onKeyDown={handleKeyDown}
-            />
-          </Grid>
-        <Grid item xs={12} md={1}>
-          <Button
-            variant='contained'
-            disabled={loading}
-            onClick={ onSend }
-            sx={{ ml: 2 }}
-          >
-            Send
-          </Button>
-        </Grid>
-      </Grid>
-    </Container>
-  ) : (<Homepage/>)
+                  sx={{mb:1}}
+                >MAKE BEAUTIFUL IMAGES</Button>
+                <br />
+                <Button
+                  variant="outlined"
+                  onClick={ () => {
+                    // endIcon={<LoginIcon />}
+                    // account.onLogin()
+                  }}
+                >FINE TUNE SDXL</Button>
+            </RightContent>
+        </Block>
+    );
+}
+
+function LanguageModelsBlock() {
+    return (
+        <Block>
+            <LeftContent>
+                <h1>Language models</h1>
+                <p>Small open source LLMs are beating proprietary models</p>
+                <Button
+                  variant="contained"
+                  onClick={ () => {
+                    // endIcon={<LoginIcon />}
+                    // account.onLogin()
+                  }}
+                  sx={{mb:1}}
+                >CHAT TO MISTRAL</Button>
+                <Button
+                  variant="outlined"
+                  onClick={ () => {
+                    // endIcon={<LoginIcon />}
+                    // account.onLogin()
+                  }}
+                >FINE TUNE MISTRAL-7B</Button>
+            </LeftContent>
+            <RightMedia>
+                <img src="/img/mistral.png" alt="Mistral-8B" style={{width:"300px", float: "right"}} />
+            </RightMedia>
+        </Block>
+    );
+}
+
+function DeploymentBlock() {
+    return (
+        <Block>
+            <LeftMedia>
+                <img src="/img/servers.png" alt="Servers in a data center" style={{width:"299px"}} />
+            </LeftMedia>
+            <RightContent>
+                <h1>Deployment</h1>
+                <ul style={{ listStyleType: 'none', padding: '0' }}>
+                    <li>GPU scheduler</li>
+                    <li>Smart runners</li>
+                    <li>Autoscaler</li>
+                </ul>
+                <Button
+                  variant="contained"
+                  onClick={ () => {
+                    // endIcon={<LoginIcon />}
+                    // account.onLogin()
+                  }}
+                  sx={{mb:1}}
+                >CONNECT RUNNER</Button>
+                <Button
+                  variant="outlined"
+                  onClick={ () => {
+                    // endIcon={<LoginIcon />}
+                    // account.onLogin()
+                  }}
+                >DEPLOY ON YOUR INFRA</Button>
+            </RightContent>
+        </Block>
+    );
+}
+
+function Footer() {
+    return (
+        <Block>
+            <LeftContent>
+                <h1>Clone us on GitHub</h1>
+                <p>Bring new models to the open stack</p>
+                <Button
+                  variant="contained"
+                  onClick={ () => {
+                    // endIcon={<LoginIcon />}
+                    // account.onLogin()
+                  }}
+                >JOIN MLOPS.COMMUNITY SLACK</Button>
+                <Button
+                  variant="outlined"
+                  onClick={ () => {
+                  }}
+                  sx={{mt:1}}
+                >GITHUB.COM/HELIX-ML/HELIX</Button>
+            </LeftContent>
+            <RightMedia>
+                <img src="/img/github.png" alt="GitHub users collaborating" style={{width:"249px", float: "right"}} />
+            </RightMedia>
+        </Block>
+    );
+}
+
+// export default App;
+const Dashboard: FC = () => {
+  return (
+      <XContainer>
+          <OpenAIBlock />
+          <ImageModelsBlock />
+          <LanguageModelsBlock />
+          <DeploymentBlock />
+          <Footer />
+      </XContainer>
+  );
 }
 
 export default Dashboard
