+++ frontend/src/router.tsx
@@ -5,6 +5,7 @@ import browserPlugin from 'router5-plugin-browser'
 import Home from './pages/Home'
 import Session from './pages/Session'
 import Account from './pages/Account'
+import New from './pages/New'
 
 import { FilestoreContextProvider } from './contexts/filestore'
 import Files from './pages/Files'
@@ -43,6 +44,15 @@ const routes: IApplicationRoute[] = [{
     <FilestoreContextProvider>
       <Files />
     </FilestoreContextProvider>
+  ),
+}, {
+  name: 'new',
+  path: '/new',
+  meta: {
+    title: 'New Session',
+  },
+  render: () => (
+      <New />
   ),
 }, {
   name: 'session',
